Refactor code for inserting upload text
